Step 1 move consul
